+++ src/main/java/com/infoshareacademy/drinkers/service/sorting/DrinkAlcoholicComparator.java
@@ -1,13 +1,13 @@
-package com.infoshareacademy.drinkers.service.filtering;
+package com.infoshareacademy.drinkers.service.sorting;
 
 import com.infoshareacademy.drinkers.domain.drink.Drink;
 
 import java.util.Comparator;
 
-public class DrinkDateModifiedComparator implements Comparator<Drink> {
+public class DrinkAlcoholicComparator implements Comparator<Drink> {
     @Override
     public int compare(Drink o1, Drink o2) {
-        return o1.getDateModified().compareTo(o2.getDateModified());
+        return o1.getAlcoholic().compareTo(o2.getAlcoholic());
     }
 
     @Override
