+++ src/main/java/com/infoshareacademy/drinkers/service/sorting/DrinkIDComparator.java
@@ -1,13 +1,13 @@
-package com.infoshareacademy.drinkers.service.filtering;
+package com.infoshareacademy.drinkers.service.sorting;
 
 import com.infoshareacademy.drinkers.domain.drink.Drink;
 
 import java.util.Comparator;
 
-public class DrinkNameComparator implements Comparator<Drink>   {
+public class DrinkIDComparator implements Comparator<Drink> {
     @Override
     public int compare(Drink o1, Drink o2) {
-        return o1.getDrink().compareTo(o2.getDrink());
+        return o1.getIdDrink().compareTo(o2.getIdDrink());
     }
 
     @Override
